# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEO & Marketing Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List keywords, optionally filtered by website
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Add a tracked keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /keywords/{id}:
    patch:
      operationId: updateKeyword
      tags: [keywords]
      summary: Update a keyword entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKeywordBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Delete a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Social Posts ─────────────────────────────────────────────────────────────
  /social-posts:
    get:
      operationId: listSocialPosts
      tags: [social-posts]
      summary: List social posts
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: createSocialPost
      tags: [social-posts]
      summary: Create a social post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialPostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
  /social-posts/{id}:
    patch:
      operationId: updateSocialPost
      tags: [social-posts]
      summary: Update a social post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSocialPostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
    delete:
      operationId: deleteSocialPost
      tags: [social-posts]
      summary: Delete a social post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Campaigns ────────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Backlinks ────────────────────────────────────────────────────────────────
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Analytics / Dashboard ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Keyword:
      type: object
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        keyword:
          type: string
        currentRank:
          type: ["integer", "null"]
        searchVolume:
          type: ["integer", "null"]
        difficulty:
          type: ["integer", "null"]
        status:
          type: string
          description: tracking | paused
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - websiteId
        - keyword
        - status
        - createdAt
        - updatedAt
    CreateKeywordBody:
      type: object
      properties:
        websiteId:
          type: integer
        keyword:
          type: string
        currentRank:
          type: ["integer", "null"]
        searchVolume:
          type: ["integer", "null"]
        difficulty:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - websiteId
        - keyword
        - status
    UpdateKeywordBody:
      type: object
      properties:
        keyword:
          type: string
        currentRank:
          type: ["integer", "null"]
        searchVolume:
          type: ["integer", "null"]
        difficulty:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]

    # ─── SocialPost ───────────────────────────────────────────────────────────
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        platform:
          type: string
          description: facebook | instagram | twitter | linkedin | youtube
        content:
          type: string
        status:
          type: string
          description: draft | scheduled | published
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        mediaUrl:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - websiteId
        - platform
        - content
        - status
        - createdAt
        - updatedAt
    CreateSocialPostBody:
      type: object
      properties:
        websiteId:
          type: integer
        platform:
          type: string
        content:
          type: string
        status:
          type: string
        scheduledAt:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]
      required:
        - websiteId
        - platform
        - content
        - status
    UpdateSocialPostBody:
      type: object
      properties:
        platform:
          type: string
        content:
          type: string
        status:
          type: string
        scheduledAt:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]

    # ─── Campaign ─────────────────────────────────────────────────────────────
    Campaign:
      type: object
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        name:
          type: string
        type:
          type: string
          description: organic | paid | email | social
        goal:
          type: string
        budget:
          type: ["number", "null"]
        status:
          type: string
          description: planning | active | paused | completed
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        spend:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - websiteId
        - name
        - type
        - goal
        - status
        - createdAt
        - updatedAt
    CreateCampaignBody:
      type: object
      properties:
        websiteId:
          type: integer
        name:
          type: string
        type:
          type: string
        goal:
          type: string
        budget:
          type: ["number", "null"]
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        spend:
          type: ["number", "null"]
      required:
        - websiteId
        - name
        - type
        - goal
        - status
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        goal:
          type: string
        budget:
          type: ["number", "null"]
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        spend:
          type: ["number", "null"]

    # ─── Backlink ─────────────────────────────────────────────────────────────
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        source:
          type: string
          description: organic | paid | social | direct | referral
        status:
          type: string
          description: new | contacted | qualified | converted | lost
        notes:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - websiteId
        - name
        - source
        - status
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        websiteId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
      required:
        - websiteId
        - name
        - source
        - status
    UpdateLeadBody:
      type: object
      properties:
        campaignId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        value:
          type: ["number", "null"]

    # ─── Analytics ────────────────────────────────────────────────────────────
...