# lib/api-zod/src/generated/types/createKeywordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SEO & Marketing Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateKeywordBody {
  websiteId: number;
  keyword: string;
  /** @nullable */
  currentRank?: number | null;
  /** @nullable */
  searchVolume?: number | null;
  /** @nullable */
  difficulty?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
}