# lib/api-zod/src/generated/types/socialPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SEO & Marketing Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SocialPost {
  id: number;
  websiteId: number;
  /** facebook | instagram | twitter | linkedin | youtube */
  platform: string;
  content: string;
  /** draft | scheduled | published */
  status: string;
  /** @nullable */
  scheduledAt?: Date | null;
  /** @nullable */
  publishedAt?: Date | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  campaignId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}