# lib/api-zod/src/generated/types/updateLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SEO & Marketing Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLeadBody {
  /** @nullable */
  campaignId?: number | null;
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  source?: string;
  status?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  value?: number | null;
}